New flag value for supporting ThreadStatic variables

+++ WifiCachingStream.cpp
@@ -135,7 +135,7 @@ void WifiCachingStream::flush()
 size_t WifiCachingStream::readBytes(char* buffer, size_t length)
 {
 	int received = 0;
-	auto result = ftp_recv_non_blocking(_connection_sd, buffer, length, &received);
+	auto result = ftp_recv_non_blocking(_connection_sd, buffer, (int)length, &received);
 	if (received >= 1)
 	{
 		return received;
